+++ src/Highmaps.php
@@ -5,7 +5,7 @@
  *
  * @author Milo Schuman <miloschuman@gmail.com>
  * @link https://github.com/miloschuman/yii2-highcharts/
- * @license http://www.opensource.org/licenses/mit-license.php MIT License
+ * @license https://www.opensource.org/licenses/mit-license.php MIT License
  */
 
 namespace miloschuman\highcharts;
@@ -16,6 +16,6 @@ namespace miloschuman\highcharts;
 class Highmaps extends Highcharts
 {
 
-    protected $constr = 'Map';
+    protected $constr = 'mapChart';
     protected $baseScript = 'modules/map';
 }